[common.tuple] refs #10

+++ dune/stuff/common/tuple.hh
@@ -213,7 +213,7 @@ namespace TupleProduct
   using boost::mpl::if_;
   using boost::mpl::deref;
 
-  unsigned int total_recursions = 0;
+  size_t total_recursions = 0;
 
   struct end_of_recursion_tag
   {
